kms-integration - moving everything to a kes file
