+++ cluster/tenants.go
@@ -17,20 +17,12 @@
 package cluster
 
 import (
-	"crypto"
-	"crypto/ed25519"
-	"crypto/rand"
-	"crypto/x509"
-	"crypto/x509/pkix"
 	"database/sql"
-	"encoding/hex"
-	"encoding/pem"
 	"errors"
 	"fmt"
 	"log"
-	"math/big"
+	"os"
 	"regexp"
-	"time"
 
 	"github.com/golang-migrate/migrate/v4"
 	"github.com/minio/minio-go/v6"
@@ -38,8 +30,6 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/client-go/informers"
-	"k8s.io/client-go/tools/cache"
 )
 
 type Tenant struct {
@@ -54,172 +44,6 @@ type AddTenantResult struct {
 	Error error
 }
 
-func CreateTenantKesSecret(secretName string, roleId string, secretId string) <-chan struct{} {
-	doneCh := make(chan struct{})
-	go func() {
-		clientSet, err := k8sClient()
-		if err != nil {
-			log.Println(err)
-			close(doneCh)
-			return
-		}
-
-		factory := informers.NewSharedInformerFactory(clientSet, 0)
-		secretInformer := factory.Core().V1().Secrets().Informer()
-		secretInformer.AddEventHandler(cache.ResourceEventHandlerFuncs{
-			AddFunc: func(obj interface{}) {
-				secret := obj.(*corev1.Secret)
-				if secret.Name == secretName {
-					log.Println(secret)
-					close(doneCh)
-					return
-				}
-			},
-		})
-
-		go secretInformer.Run(doneCh)
-
-		_, err = clientSet.CoreV1().Secrets("default").Create(&corev1.Secret{
-			ObjectMeta: metav1.ObjectMeta{
-				Name: secretName,
-				Labels: map[string]string{
-					"name": secretName,
-				},
-			},
-			StringData: map[string]string{
-				"role_id":   roleId,
-				"secret_id": secretId,
-			},
-		})
-		if err != nil {
-			log.Println(err)
-			close(doneCh)
-		}
-	}()
-	return doneCh
-}
-
-func deployNewKes() <-chan struct{} {
-	doneCh := make(chan struct{})
-	go func() {
-		close(doneCh)
-	}()
-	return doneCh
-}
-
-type KeyPair struct {
-	cert         string
-	certIdentity string
-	key          string
-}
-
-func generateKeyPair(name string) chan *KeyPair {
-	doneCh := make(chan *KeyPair)
-	go func() {
-		defer close(doneCh)
-		public, private, err := ed25519.GenerateKey(rand.Reader)
-		if err != nil {
-			log.Println(err)
-			return
-		}
-		serialNumberLimit := new(big.Int).Lsh(big.NewInt(1), 128)
-		serialNumber, err := rand.Int(rand.Reader, serialNumberLimit)
-		if err != nil {
-			log.Println(err)
-			return
-		}
-		now := time.Now()
-		template := x509.Certificate{
-			SerialNumber: serialNumber,
-			Subject: pkix.Name{
-				CommonName: name,
-			},
-			NotBefore:             now,
-			NotAfter:              now.Add(87660 * time.Hour),
-			KeyUsage:              x509.KeyUsageDigitalSignature,
-			ExtKeyUsage:           []x509.ExtKeyUsage{x509.ExtKeyUsageClientAuth},
-			BasicConstraintsValid: true,
-		}
-
-		derBytes, err := x509.CreateCertificate(rand.Reader, &template, &template, public, private)
-		if err != nil {
-			log.Println(err)
-			return
-		}
-		privBytes, err := x509.MarshalPKCS8PrivateKey(private)
-		if err != nil {
-			log.Println(err)
-			return
-		}
-		key := string(pem.EncodeToMemory(&pem.Block{Type: "PRIVATE KEY", Bytes: privBytes}))
-		cert := string(pem.EncodeToMemory(&pem.Block{Type: "CERTIFICATE", Bytes: derBytes}))
-
-		h := crypto.SHA256.New()
-		publicCertificate, _ := x509.ParseCertificate(derBytes)
-		h.Write(publicCertificate.RawSubjectPublicKeyInfo)
-
-		doneCh <- &KeyPair{
-			key:          key,
-			cert:         cert,
-			certIdentity: hex.EncodeToString(h.Sum(nil)),
-		}
-	}()
-	return doneCh
-}
-
-func storeKeyPairInSecret(secretName string, content map[string]string) chan struct{} {
-	doneCh := make(chan struct{})
-	go func() {
-		defer close(doneCh)
-		clientSet, err := k8sClient()
-		if err != nil {
-			log.Println(err)
-			return
-		}
-		factory := informers.NewSharedInformerFactory(clientSet, 0)
-		secretInformer := factory.Core().V1().Secrets().Informer()
-		secretInformer.AddEventHandler(cache.ResourceEventHandlerFuncs{
-			AddFunc: func(obj interface{}) {
-				secret := obj.(*corev1.Secret)
-				if secret.Name == secretName {
-					log.Println(secret)
-					return
-				}
-			},
-		})
-
-		go secretInformer.Run(doneCh)
-
-		_, err = clientSet.CoreV1().Secrets("default").Create(&corev1.Secret{
-			ObjectMeta: metav1.ObjectMeta{
-				Name: secretName,
-				Labels: map[string]string{
-					"name": secretName,
-				},
-			},
-			StringData: content,
-		})
-		if err != nil {
-			log.Println(err)
-			close(doneCh)
-		}
-
-	}()
-	return doneCh
-}
-
-func GenerateKeyPairAndStoreInSecret(name string) *KeyPair {
-	kesKeyPair := <-generateKeyPair(name)
-	if kesKeyPair != nil && kesKeyPair.cert != "" && kesKeyPair.key != "" {
-		<-storeKeyPairInSecret(name, map[string]string{
-			"cert":         kesKeyPair.cert,
-			"key":          kesKeyPair.key,
-			"certIdentity": kesKeyPair.certIdentity,
-		})
-	}
-	return kesKeyPair
-}
-
 // TenantAddAction adds a tenant to the cluster, if an admin name and email are provided, the user is created and invited
 // via email.
 func TenantAddAction(ctx *Context, name, shortName, userName, userEmail string) error {
@@ -232,103 +56,9 @@ func TenantAddAction(ctx *Context, name, shortName, userName, userEmail string) 
 	if !available {
 		return errors.New("Error tenant's shortname not available")
 	}
-	if GetInstance().KmsClient != nil {
-		kms := GetInstance().KmsClient
-		policyName := fmt.Sprintf("%s-kes-policy", shortName)
-
-		existingPolicy, _ := kms.Sys().GetPolicy(policyName)
-
-		if existingPolicy != "" {
-			log.Println("Error creating policy on external kms because the policy already exists ", policyName, existingPolicy)
-			return errors.New("Error adding the tenant to the db")
-		}
-
-		policyRules := fmt.Sprintf(`
-		path "kv/%s/*" {
-				capabilities = [ "create", "read", "delete" ]
-		}
-		`, shortName)
-
-		err := kms.Sys().PutPolicy(policyName, policyRules)
-		if err != nil {
-			log.Println("Error creating policy on external kms: ", err)
-			return errors.New("Error adding the tenant to the db")
-		}
-		data := map[string]interface{}{
-			"policy":             policyName,
-			"token_num_uses":     0,
-			"secret_id_num_uses": 0,
-			"period":             "5m",
-		}
-		roleName := fmt.Sprintf("%s-kes-role", shortName)
-		_, err = kms.Logical().Write(fmt.Sprintf("auth/approle/role/%s", roleName), data)
-		if err != nil {
-			log.Println("Error creating new role on external kms: ", shortName, data, err)
-			return errors.New("Error adding the tenant to the db")
-		}
-		role, err := kms.Logical().Read(fmt.Sprintf("auth/approle/role/%s/role-id", roleName))
-		if err != nil {
-			log.Println("Error reading role_id from external kms: ", shortName, err)
-			return errors.New("Error adding the tenant to the db")
-		}
-		roleSecret, err := kms.Logical().Write(fmt.Sprintf("auth/approle/role/%s/secret-id", roleName), map[string]interface{}{})
-		if err != nil {
-			log.Println("Error reading role_secret_id from external kms: ", shortName)
-			return errors.New("Error adding the tenant to the db")
-		}
-
-		roleId := role.Data["role_id"].(string)
-		roleSecretId := roleSecret.Data["secret_id"].(string)
-
-		//tenantKmsSecrets := fmt.Sprintf("%s-kms-credentials", shortName)
-		//<-CreateTenantKesSecret(tenantKmsSecrets, roleId, roleSecretId)
-
-		kesServerKeyPair := fmt.Sprintf("%s-kes-server-keypair", shortName)
-		kesAppKeyPair := fmt.Sprintf("%s-kes-app-keypair", shortName)
-		appKeys := GenerateKeyPairAndStoreInSecret(kesAppKeyPair)
-		GenerateKeyPairAndStoreInSecret(kesServerKeyPair)
-
-		kesServerConfig := fmt.Sprintf(`
-			address = "127.0.0.1:7373"
-			root = "disabled"
-			
-			[tls]
-			key  = "kes-config/server/key"
-			cert = "kes-config/server/cert"
-			
-			[policy.prod-app]
-			paths      = [ "/v1/key/create/app-key", "/v1/key/generate/app-key" , "/v1/key/decrypt/app-key"]
-			identities = [ "%s" ]
-			
-			[keystore.vault]
-			address = "%s"
-			[keystore.vault.approle]
-			id     = "%s"
-			secret = "%s"
-			retry  = "15s"
-			[keystore.vault.status]
-			ping = "10s"
-		`, appKeys.certIdentity, kms.Address(), roleId, roleSecretId)
-
-		kesServerConfigSecret := fmt.Sprintf("%s-kes-server-config", shortName)
-		<-storeKeyPairInSecret(kesServerConfigSecret, map[string]string{
-			"server-config.toml": kesServerConfig,
-		})
-
-		//<-deployNewKes()
-
-		//generate secret server.key and server.cert for kes
-		//generate secret root.key and root.cert for kes
-		//generate secret app.key and app.cert for kes
-
-		//calculate the identity for root.cert
-		//calculate the identity for app.cert
-
-		//generate toml configmap for kes using previous values
-		//deploy new kes and mount the server-config.toml configmap
-		//deploy new minio tenant and mount app.key, app.cert, server.cert
+	if os.Getenv("KMS_ADDRESS") != "" {
+		<-StartNewKes(shortName)
 	}
-
 	//// first find a cluster where to allocate the tenant
 	//sg := <-SelectSGWithSpace(ctx)
 	//if sg.Error != nil {
